requirements/dev.txt
-r base.txt
django-debug-toolbar==1.3.2
# Required by django-debug-toolbar
sqlparse==0.1.16

coverage==3.7.1
pylint==1.1.0
# Required by pylint
logilab-common==0.61.0
astroid==1.0.1
coveralls==0.4.2
flake8==2.4.1
# Required by flake8
pyflakes==0.8.1
pep8==1.5.7
mccabe==0.3.1

# For translation
transifex-client==0.11b3

PyYAML==3.10
Sphinx==1.2.3
django-extensions==1.5.7
